test/python-kiota/pyproject.toml
[tool.poetry]
name = "modelregistrysdk"
version = "1.0.0"
description = ""
authors = ["Andrea Peruffo <andrea.peruffo1982@gmail.com>"]
readme = "README.md"
packages = [{include = "apisdk"}]
include = [
    { path = "openapi.yaml", format=["sdist", "wheel"] },
    { path = "apisdk/client/**/*", format=["sdist", "wheel"] },
]
exclude = [
    { path = "apisdk/__pycache__", format=["sdist", "wheel"] },
]
license = "Apache 2.0"
homepage = "https://github.com/kubeflow/model-registry"
repository = "https://github.com/kubeflow/model-registry"
keywords = ["model", "registry"]

[tool.poetry.dependencies]
python = "^3.9"
microsoft-kiota-abstractions = "^1.3.2"
microsoft-kiota-http = "^1.3.1"
microsoft-kiota-serialization-json = "^1.1.0"
microsoft-kiota-serialization-text = "^1.0.0"
microsoft-kiota-serialization-form = "^0.1.0"
microsoft-kiota-serialization-multipart = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
requests = "2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core", "requests", "shutils"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
generate-setup-file = false
script = "kiota-gen.py"

[tool.pytest.ini_options]
pythonpath = [ "apisdk" ]
asyncio_mode = "auto"

[tool.black]
extend-exclude = 'client'
